Rework DD classes
- Remove lots of superfluous layers of inheritance. Merge
  AbstractDD into DD, merge AbstractParallelDD and
  CombinedParallelDD into ParallelDD.

- Unify iterators in sequential and parallel DD. Now, a config
  iterator has to return numbers in the -n .. +n-1 range, 0 .. +n-1
  denoting subsets and -n .. -1 denoting complements.
  CombinedIterator still allows to compose such a new-style
  iterator from a pair of classic subset iterator and complement
  iterator.
